drivers: Fix LG's inclusion of MobiCore driver
It seems the logic is as follows:
- KR region devices *must* build MobiCore driver
- style3lm *must* build MobiCore driver
- caymanslm *must not* build MobiCore driver

CONFIG_TRUSTONIC_TEE defaults to `y`. So, instead of this logic, just
include the directory anyway, and disable such devices that do not
build this driver from defconfig.

judy* targets do not build this driver, as the defconfig used is based
on the global, non-KR target.
caymanslm does not build this driver as it was specifically excluded in
LG's Makefile changes.

Change-Id: Ie80ec5207500ddfb5c3d8bc2eea838c0e9089de4
Signed-off-by: Han Sol Jin <hansol@hansol.ca>