msm: cam_actuator: fix possibility of actuator_tele sysfs being created multiple times
The previous code checked whether a given actuator needed a sysfs
created. If 0, it would create `actuator_normal`, in all other cases it
created `actuator_tele`.

Somewhere outside of our control, this caused the `else` condition to
be triggered multiple times, leading to a null dereference.

Checking devices that actually use this sysfs (judypn, caymanslm,
style3lm), we can observe that, for the `else` scenario, there is only
`qcom,actuator@3`, present in judypn's devicetree. So, change the code
to explicitly check for this.

Change-Id: Ib5c16a71b9ca962eb1f97bb99667f17098061685
Signed-off-by: Han Sol Jin <hansol@hansol.ca>